requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.5
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
stripe==5.0.0
tzdata==2022.6
urllib3==1.26.12
gunicorn==20.1.0
dj-database-url==0.5.0